chore: reconfigure test workspace
Previously, the workspace contained both children packages used for e2e
tests as well as packages that actually formulate the base of the plugin.

This created much confusion for me as I had to pre-build the composer plugin
in order to do anything else. By splitting the workspace apart, I get to avoid
this problem. With any luck, I'll be able to use the new `testworkspace`
in a separate e2e test.